requirements.txt
#api framework
fastapi

#ASGI server
uvicorn

#templating lib for python
fastapi_chameleon

#serving static
aiofiles

#orm
sqlalchemy

#async driver for DB
asyncpg
#sync driver for DB
psycopg2

#storing env
python-dotenv

#password handling
passlib

#for form parsing
python-multipart

#for JWT tokens
python-jose